Split out preprocess module
