+++ src/lib.rs
@@ -6,6 +6,7 @@ extern crate failure;
 extern crate lazy_static;
 
 mod syntax;
+mod preprocess;
 
 use walkdir::WalkDir;
 
@@ -14,14 +15,17 @@ use failure::ResultExt;
 
 use std::collections::BTreeMap as Map;
 use std::collections::BTreeSet as Set;
-use std::fs;
 use std::path::{Path, PathBuf};
+
+use preprocess::Parsed;
 
 pub fn run(root_path: impl AsRef<Path>, emitter: &mut Emitter) -> Result<(), Error> {
     run_(root_path.as_ref(), emitter)
 }
 
 fn run_(root_path: &Path, emitter: &mut Emitter) -> Result<(), Error> {
+    use preprocess::PreprocessOutput;
+
     let mut files = Map::new();
 
     for entry in WalkDir::new(root_path) {
@@ -36,7 +40,7 @@ fn run_(root_path: &Path, emitter: &mut Emitter) -> Result<(), Error> {
             continue;
         }
 
-        match parse_and_preprocess(entry.path(), emitter)? {
+        match preprocess::parse_and_preprocess(entry.path(), emitter)? {
             PreprocessOutput::Valid(parsed) => {
                 let path = entry.path().canonicalize()?;
                 files.insert(path, parsed);
@@ -137,125 +141,6 @@ impl Emitter for VecEmitter {
         let to_emit = EmittedItem { kind, message, location, notes };
         self.emitted_items.push(to_emit)
     }
-}
-
-// ---- Preprocessing --------------------------------------------------------
-
-/// Parsed and preprocessed source file
-#[derive(Debug)]
-struct Parsed {
-    imports: Vec<PathBuf>,
-    definitions: Vec<syntax::Definition>,
-    usages: Vec<syntax::Usage>,
-
-    /// Original, non-resolved path, relative to PWD
-    original_path: PathBuf,
-}
-
-#[derive(Debug)]
-enum PreprocessOutput {
-    /// Parsed and preprocessed file
-    Valid(Parsed),
-
-    /// A file can't be preprocessed since it contains invalid imports
-    InvalidImports,
-}
-
-/// Parses and preprocesses a file for further analysys.
-fn parse_and_preprocess(path: &Path, emitter: &mut Emitter) -> Result<PreprocessOutput, Error> {
-    let source = fs::read_to_string(path)?;
-    let file = syntax::parse(&source);
-
-    for testcase in file.testcases {
-        let invalid_chars: &[char] = &['"', '>', '<', '|', ':', '*', '?', '\\', '/'];
-
-        if file.uses_pester_logger && testcase.name.contains(invalid_chars) {
-            emitter.emit(
-                Message::Warning,
-                "Testname contains invalid characters".to_owned(),
-                testcase.location.in_file(path),
-                Some(format!(
-                    "These characters are invalid in a file name: {:?}",
-                    invalid_chars,
-                )),
-            );
-        }
-    }
-
-    let resolved_imports = match resolve_imports(path, file.imports, emitter)? {
-        Some(imports) => imports,
-        None => return Ok(PreprocessOutput::InvalidImports),
-    };
-
-    Ok(PreprocessOutput::Valid(Parsed {
-        imports: resolved_imports,
-        definitions: file.definitions,
-        usages: file.usages,
-        original_path: path.to_owned(),
-    }))
-}
-
-/// Verifies imports and canonicalizes their paths
-///
-/// Returns None if any of imports were not recognized
-fn resolve_imports(source_path: &Path, imports: Vec<syntax::Import>, emitter: &mut Emitter)
-    -> Result<Option<Vec<PathBuf>>, Error>
-{
-    let mut import_error = false;
-    let mut resolved_imports = Vec::new();
-
-    for import in imports {
-        use syntax::Importee;
-
-        let dir = source_path.parent().unwrap();
-        let filename = source_path.file_name().unwrap().to_str().unwrap();
-
-        let dest_path = match import.importee {
-            Importee::Relative(relative_path) => dir.join(relative_path),
-            Importee::HereSut => dir.join(filename.replace(".Tests", "")),
-            Importee::Unrecognized(_) => {
-                // Should we treat unrecognized import as an error also?
-                // This will stop processing the file further and will result in
-                // less spammy output, because we'll probably get some
-                // "Not in scope" errors later on.
-                // import_error = true;
-
-                emitter.emit(
-                    Message::Warning,
-                    "Unrecognized import statement".to_string(),
-                    import.location.in_file(source_path),
-                    Some(
-                        "Note: Recognized imports are `$PSScriptRoot\\..` or `$here\\$sut`"
-                            .to_string(),
-                    ),
-                );
-
-                continue;
-            }
-        };
-
-        if dest_path.exists() {
-            resolved_imports.push(dest_path.canonicalize()?)
-        } else {
-            import_error = true;
-
-            emitter.emit(
-                Message::Error,
-                "Invalid import".to_string(),
-                import.location.in_file(source_path),
-                Some(format!(
-                    "File not found: {}",
-                    dest_path.display()
-                )),
-            );
-        }
-    }
-
-    if import_error {
-        return Ok(None)
-    }
-
-    Ok(Some(resolved_imports))
 }
 
 // ---- Scope analysis -------------------------------------------------------
